Anul II/Semestrul II/MDS/ProiectMDS/calorie-tracker-backend/app/apis/schemas/errors_schema.py
from pydantic import BaseModel


class ApiError(BaseModel):
    detail: str

    class Config:
        schema_extra = {
            "example": {
                "detail": "<error description>",
            }
        }
